% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{make_umap_object}
\alias{make_umap_object}
\title{Title}
\usage{
make_umap_object(umap_result, annotation = NULL)
}
\arguments{
\item{umap_result}{- output of running}

\item{annotation}{- optional annotation file}
}
\value{
- a umap object that includes plotting
}
\description{
Title
}
\examples{

library(flowCore)

umap_table <- umap()
}
