% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_app.R
\name{run_umap_shiny}
\alias{run_umap_shiny}
\title{Open a shiny app to explore the data in a UMAP embedding.}
\usage{
run_umap_shiny(umap)
}
\arguments{
\item{umap}{output of a call to `umap`}
}
\value{
Open an interactive shiny app to explore the data.
}
\description{
Open a shiny app to explore the data in a UMAP embedding.
}
