% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{umap}
\usage{
umap(data, ...)
}
\arguments{
\item{data}{input data matrix}

\item{...}{arguments to umap object}
}
\value{
matrix
}
\description{
umap
}
\examples{
umap(as.matrix(iris[, 1:4]))
}
